convert csv to excel in python
oad the CSV as a pandas data frame. Convert the CSV file to an excel file by passing the index as True as an argument to the excel() function and displaying index values. index=True means that index values are shown here. Use the save() function (saves the file) to save the result/output excel file.